Replace Toast by SnackBar:
it seems the Snackbar uses a dark theme on light app theme and vise versa.
I don't like this at all, so I changed the Snackbar theme to match the app theme.
When using a Snackbar, the app theme needs to inherit from a Material Components theme,
which causes a lot of work for updating the app theme, and replace AppCompat components by
Material Components.
(At least to much work for only using a Snackbar).
Fortunately, there are Material Components Bridge themes,
which can be used without a lot of work.